feat: GlobalExceptionHanlder 도입
- ErrorCode 및 ErrorResponse 도입
- Auth 패키지 내 AuthorizationException 도입
- GlobalExceptionHanlder 내 Auth 패키지 Exception 처리 추가

+++ src/main/java/nextstep/schedule/ScheduleService.java
@@ -1,5 +1,12 @@
 package nextstep.schedule;
 
+import java.util.Objects;
+import nextstep.error.ErrorCode;
+import nextstep.error.exception.RoomReservationException;
+import nextstep.reservation.dao.ReservationDao;
+import nextstep.reservation.dao.ReservationWaitingDao;
+import nextstep.reservation.domain.Reservation;
+import nextstep.reservation.domain.ReservationWaiting;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
@@ -10,10 +17,15 @@ import java.util.List;
 public class ScheduleService {
     private ScheduleDao scheduleDao;
     private ThemeDao themeDao;
+    private ReservationDao reservationDao;
+    private ReservationWaitingDao reservationWaitingDao;
 
-    public ScheduleService(ScheduleDao scheduleDao, ThemeDao themeDao) {
+    public ScheduleService(ScheduleDao scheduleDao, ThemeDao themeDao, ReservationDao reservationDao,
+                           ReservationWaitingDao reservationWaitingDao) {
         this.scheduleDao = scheduleDao;
         this.themeDao = themeDao;
+        this.reservationDao = reservationDao;
+        this.reservationWaitingDao = reservationWaitingDao;
     }
 
     public Long create(ScheduleRequest scheduleRequest) {
@@ -26,6 +38,15 @@ public class ScheduleService {
     }
 
     public void deleteById(Long id) {
+        Schedule schedule = scheduleDao.findById(id);
+        if(Objects.isNull(schedule)) {
+            throw new RoomReservationException(ErrorCode.SCHEDULE_NOT_FOUND);
+        }
+        List<Reservation> reservationList = reservationDao.findByScheduleId(id);
+        List<ReservationWaiting> reservationWaitingList = reservationWaitingDao.findByScheduleId(id);
+        if (reservationList.size() > 0 || reservationWaitingList.size() > 0) {
+            throw new RoomReservationException(ErrorCode.SCHEDULE_CANT_BE_DELETED);
+        }
         scheduleDao.deleteById(id);
     }
 }
