+++ src/main/java/nextstep/theme/ThemeService.java
@@ -1,5 +1,9 @@
 package nextstep.theme;
 
+import nextstep.error.ErrorCode;
+import nextstep.error.exception.RoomReservationException;
+import nextstep.schedule.Schedule;
+import nextstep.schedule.ScheduleDao;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -7,9 +11,11 @@ import java.util.List;
 @Service
 public class ThemeService {
     private ThemeDao themeDao;
+    private ScheduleDao scheduleDao;
 
-    public ThemeService(ThemeDao themeDao) {
+    public ThemeService(ThemeDao themeDao, ScheduleDao scheduleDao) {
         this.themeDao = themeDao;
+        this.scheduleDao = scheduleDao;
     }
 
     public Long create(ThemeRequest themeRequest) {
@@ -23,9 +29,12 @@ public class ThemeService {
     public void delete(Long id) {
         Theme theme = themeDao.findById(id);
         if (theme == null) {
-            throw new NullPointerException();
+            throw new RoomReservationException(ErrorCode.THEME_NOT_FOUND);
         }
-
+        List<Schedule> schedules = scheduleDao.findByThemeId(id);
+        if (schedules.size() > 0) {
+            throw new RoomReservationException(ErrorCode.THEME_CANT_BE_DELETED);
+        }
         themeDao.delete(id);
     }
 }
